# improve performance of findValues for common tags